StepTwoForm component using functions from service.js and getUserLocation files (refactor, less code used).
